+++ packages/rc-components/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useLocationQuery';
 export * from './useDefaultProConfig';
 export * from './utils';
+export * from './useToolbarMenuItems'