+++ packages/service-api/src/scm.ts
@@ -134,20 +134,21 @@ export const scmPostMaterialsTypeTable = ({
   );
 };
 
-
 // /v1/common/generateShareCode
 
-export type IScmPostGenerateShareCodeProps =   Partial<{
-  busCode:string;
-  busId:string;
-  companyId:string;
-  companyName:string;
-  otherInfos:any;
-}>
+export type IScmPostGenerateShareCodeProps = Partial<{
+  busCode: string;
+  busId: string;
+  name?: string;
+  orderTypeName?: string;
+  companyId: string;
+  companyName: string;
+  otherInfos: any;
+}>;
 
-export interface IScmPostGenerateShareCodeData{
+export interface IScmPostGenerateShareCodeData {
   shareCode: string;
-        shareUrl: string;
+  shareUrl: string;
 }
 
 export const scmPostGenerateShareCode = ({