+++ packages/auth-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @core/auth-sdk
 
+## 1.1.23
+
+### Patch Changes
+
+- feat: 修复 activekey
+- Updated dependencies
+  - @core/service-api@2.3.23
+  - @core/shared@2.3.11
+
+## 1.1.22
+
+### Patch Changes
+
+- toolbar 的 hooks
+- Updated dependencies
+  - @core/service-api@2.3.22
+  - @core/shared@2.3.10
+
 ## 1.1.21
 
 ### Patch Changes