+++ packages/minio-sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 2.3.23
+
+### Patch Changes
+
+- feat: 修复 activekey
+- Updated dependencies
+  - @core/service-api@2.3.23
+
+## 2.3.22
+
+### Patch Changes
+
+- toolbar 的 hooks
+- Updated dependencies
+  - @core/service-api@2.3.22
+
 ## 2.3.21
 
 ### Patch Changes