+++ packages/service-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @core/service-api
 
+## 2.3.23
+
+### Patch Changes
+
+- feat: 修复 activekey
+
+## 2.3.22
+
+### Patch Changes
+
+- toolbar 的 hooks
+
 ## 2.3.21
 
 ### Patch Changes